:art: TCAD code restructure
