+++ tidy3d/__init__.py
@@ -1,53 +1,56 @@
 """Tidy3d package imports"""
 
-# grid
-# apodization
-# heat
-# heat
-from tidy3d.components.material.tcad.heat import (
+from tidy3d.components.material.multi_physics import MultiPhysicsMedium
+from tidy3d.components.material.tcad.charge import (
     ChargeConductorMedium,
     ChargeInsulatorMedium,
-    FluidSpec,
     SemiconductorMedium,
+)
+from tidy3d.components.material.tcad.heat import (
+    FluidSpec,
     SolidSpec,
 )
-from tidy3d.components.tcad.boundary.heat import (
-    ConvectionBC,
-    CurrentBC,
+from tidy3d.components.spice.analysis.dc import TransferFunctionDC
+from tidy3d.components.spice.sources.dc import DCTransferSource, MultiDCTransferSource
+from tidy3d.components.tcad.boundary.specification import (
     HeatBoundarySpec,
     HeatChargeBoundarySpec,
-    HeatFluxBC,
-    InsulatingBC,
-    TemperatureBC,
-    VoltageBC,
 )
-from tidy3d.components.tcad.data.monitor_data.monitor_data import (
+from tidy3d.components.tcad.data.sim_data import HeatChargeSimulationData, HeatSimulationData
+from tidy3d.components.tcad.data.types import (
     SteadyCapacitanceData,
     SteadyFreeCarrierData,
     SteadyPotentialData,
     SteadyVoltageData,
     TemperatureData,
 )
-from tidy3d.components.tcad.data.sim_data import HeatChargeSimulationData, HeatSimulationData
 from tidy3d.components.tcad.doping import ConstantDoping, GaussianDoping
 from tidy3d.components.tcad.grid import DistanceUnstructuredGrid, UniformUnstructuredGrid
-from tidy3d.components.tcad.monitors.heat import (
+from tidy3d.components.tcad.monitors.charge import (
     SteadyCapacitanceMonitor,
     SteadyFreeChargeCarrierMonitor,
     SteadyVoltageMonitor,
+)
+from tidy3d.components.tcad.monitors.heat import (
     TemperatureMonitor,
 )
 from tidy3d.components.tcad.simulation.heat import HeatSimulation
 from tidy3d.components.tcad.simulation.heat_charge import HeatChargeSimulation
-from tidy3d.components.tcad.source.heat import HeatFromElectricSource, HeatSource, UniformHeatSource
 from tidy3d.components.tcad.types import (
     AugerRecombination,
     CaugheyThomasMobility,
-    ChargeToleranceSpec,
-    DCSpec,
+    ConvectionBC,
+    CurrentBC,
+    HeatFluxBC,
+    HeatFromElectricSource,
+    HeatSource,
+    InsulatingBC,
     RadiativeRecombination,
     ShockleyReedHallRecombination,
     SlotboomNarrowingBandGap,
+    TemperatureBC,
+    UniformHeatSource,
+    VoltageBC,
 )
 
 from .components.apodization import ApodizationSpec
@@ -556,8 +559,6 @@ __all__ = [
     "SteadyPotentialData",
     "SteadyFreeCarrierData",
     "SteadyCapacitanceData",
-    "ChargeToleranceSpec",
-    "DCSpec",
     "CaugheyThomasMobility",
     "SlotboomNarrowingBandGap",
     "ShockleyReedHallRecombination",
@@ -619,4 +620,8 @@ __all__ = [
     "EMEFreqSweep",
     "FixedAngleSpec",
     "FixedInPlaneKSpec",
+    "MultiPhysicsMedium",
+    "DCTransferSource",
+    "MultiDCTransferSource",
+    "TransferFunctionDC",
 ]
