+++ tidy3d/components/material/multi_physics.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from tidy3d.components.base import Tidy3dBaseModel
 from tidy3d.components.material.solver_types import (
     ChargeMediumTypes,
@@ -12,7 +14,13 @@ class MultiPhysicsMedium(Tidy3dBaseModel):
     A multi-physics medium may contain multiple multi-physical properties, defined for each solver medium.
     """
 
-    optical: OpticalMediumTypes
-    electrical: ElectricalMediumTypes
-    heat: HeatMediumTypes
-    charge: ChargeMediumTypes
+    # TODO requires backwards compatibility.
+    name: Optional[str]
+    optical: Optional[OpticalMediumTypes]
+    electrical: Optional[ElectricalMediumTypes]
+    heat: Optional[HeatMediumTypes]
+    charge: Optional[ChargeMediumTypes]
+
+    @property
+    def heat_spec(self):
+        return self.heat
