+++ tidy3d/components/spice/analysis/dc.py
@@ -2,18 +2,14 @@
 This class defines standard SPICE electrical_analysis types (electrical simulations configurations).
 """
 
-from typing import Optional
+from typing import Optional, Tuple, Union
 
 import pydantic.v1 as pd
 
 from tidy3d.components.base import Tidy3dBaseModel
-from tidy3d.components.spice.sources.dc import MultiStaticTransferSourceDC, StaticTransferSourceDC
-
-
-class OperatingPointDC(Tidy3dBaseModel):
-    """
-    Equivalent to Section 11.1.2 in the ngspice manual.
-    """
+from tidy3d.components.spice.sources.dc import DCTransferSource, MultiDCTransferSource
+from tidy3d.components.tcad.types import HeatChargeMonitorTypes
+from tidy3d.components.types import annotate_type
 
 
 class TransferFunctionDC(Tidy3dBaseModel):
@@ -35,12 +31,16 @@ class TransferFunctionDC(Tidy3dBaseModel):
     >>> charge_settings = td.ChargeToleranceSpec(abs_tol=1e8, rel_tol=1e-10, max_iters=30)
     """
 
-    sources: StaticTransferSourceDC | MultiStaticTransferSourceDC = []
+    input: Union[MultiDCTransferSource, DCTransferSource] = []  # todo accept a single source
+    output: Tuple[
+        annotate_type(HeatChargeMonitorTypes), ...
+    ] = ()  # TODO this should be more generic, # TODO this should be a separate generic monitor class.
 
     absolute_tolerance: Optional[pd.PositiveFloat] = pd.Field(
         default=1e10,
         title="Absolute tolerance.",
-        description="Absolute tolerance used as stop criteria when converging towards a solution. Should be equivalent to the"
+        description="Absolute tolerance used as stop criteria when converging towards a solution. Should be "
+        "equivalent to the"
         "SPICE ABSTOL DC transfer parameter. TODO MARC units, TODO check equivalence. TODO what does this mean?",
     )
