+++ tidy3d/components/tcad/source/coupled.py
@@ -1,0 +1,17 @@
+"""Defines heat-charge material specifications for 'HeatChargeSimulation'"""
+
+from __future__ import annotations
+
+from tidy3d.components.tcad.source.abstract import GlobalHeatChargeSource
+
+
+class HeatFromElectricSource(GlobalHeatChargeSource):
+    """Volumetric heat source generated from an electric simulation.
+    If a `HeatFromElectricSource` is specified as a source, appropriate boundary
+    conditions for an electric simulation must be provided, since such a simulation
+    will be executed before the heat simulation can run.
+
+    Example
+    -------
+    >>> heat_source = HeatFromElectricSource()
+    """
